+++ src/hooks/ssh/SshHook.hpp
@@ -29,31 +29,38 @@ class SshHook {
 public:
     void generateSshKeys(bool overwriteSshKeysIfExist);
     void checkUserHasSshKeys();
-    void startSshd();
+    void startSshDaemon();
 
 private:
     void parseConfigJSONOfBundle();
     bool userHasSshKeys() const;
-    boost::filesystem::path getSshKeysDir() const;
+    std::string getUsername(uid_t) const;
+    boost::filesystem::path getSshKeysDirInHost(const std::string& username) const;
     void sshKeygen(const boost::filesystem::path& outputFile) const;
-    void checkThatOpenSshIsUntamperable() const;
-    void copyKeysIntoBundle() const;
-    void bindMountSshBinary() const;
+    void generateAuthorizedKeys(const boost::filesystem::path& userKeyFile,
+                                const boost::filesystem::path& authorizedKeysFile) const;
+    void copyDropbearIntoContainer() const;
+    void setupSshKeysDirInContainer() const;
+    void copySshKeysIntoContainer() const;
+    void createSshExecutableInContainer() const;
     void patchPasswdIfNecessary() const;
-    void startSshdInContainer() const;
+    void startSshDaemonInContainer() const;
     void log(const boost::format& message, sarus::common::LogLevel level) const;
     void log(const std::string& message, sarus::common::LogLevel level) const;
 
 private:
     bool isHookEnabled = false;
-    boost::filesystem::path sshKeysDir;
-    boost::filesystem::path opensshDirInHost;
-    boost::filesystem::path opensshDirInBundle;
+    std::string username;
+    boost::filesystem::path sshKeysDirInHost;
+    boost::filesystem::path sshKeysDirInContainer;
+    boost::filesystem::path dropbearDirInHost;
+    boost::filesystem::path dropbearDirInContainer;
     boost::filesystem::path bundleDir;
     boost::filesystem::path rootfsDir;
     pid_t pidOfContainer;
     uid_t uidOfUser;
     gid_t gidOfUser;
+    std::uint16_t serverPort;
 };
 
 }}} // namespace